Added middleware control

+++ api/routes/api.php
@@ -29,6 +29,8 @@ Route::prefix(AuthEnum::ROUTE_PREFIX)->group(function () {
 	Route::get('/logout', [AuthController::class, 'logout'])->name('auth.logout');
 });
 
+Route::post(UserEnum::ROUTE_PREFIX . '/', [UserController::class, 'store'])->name('users.store');
+
 Route::middleware([JwtMiddleware::class])->group(function () {
 	/**
 	 * USERS
@@ -36,7 +38,6 @@ Route::middleware([JwtMiddleware::class])->group(function () {
 	Route::prefix(UserEnum::ROUTE_PREFIX)->group(function () {
 		Route::get('/', [UserController::class, 'index'])->name('users.index')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
 		Route::get('/{id}', [UserController::class, 'show'])->name('users.show');
-		Route::post('/', [UserController::class, 'store'])->name('users.store');
 		Route::post('/import/leads', [UserController::class, 'importLeads'])->name('users.importLeads')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN);
 		Route::put('/{id}', [UserController::class, 'update'])->name('users.update');
 		Route::delete('/{id}', [UserController::class, 'destroy'])->name('users.delete');
@@ -46,11 +47,11 @@ Route::middleware([JwtMiddleware::class])->group(function () {
 	 * ASSISTANCES
 	 */
 	Route::prefix(AssistanceEnum::ROUTE_PREFIX)->group(function () {
-		Route::get('/', [AssistanceController::class, 'index'])->name('assistances.index');
-		Route::get('/{id}', [AssistanceController::class, 'show'])->name('assistances.show');
+		Route::get('/', [AssistanceController::class, 'index'])->name('assistances.index')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::get('/{id}', [AssistanceController::class, 'show'])->name('assistances.show')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
 		Route::post('/', [AssistanceController::class, 'store'])->name('assistances.store');
-		Route::put('/{id}', [AssistanceController::class, 'update'])->name('assistances.update');
-		Route::delete('/{id}', [AssistanceController::class, 'destroy'])->name('assistances.delete');
+		Route::put('/{id}', [AssistanceController::class, 'update'])->name('assistances.update')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::delete('/{id}', [AssistanceController::class, 'destroy'])->name('assistances.delete')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
 	});
 
 	/**
@@ -58,10 +59,10 @@ Route::middleware([JwtMiddleware::class])->group(function () {
 	 */
 	Route::prefix(GenreEnum::ROUTE_PREFIX)->group(function () {
 		Route::get('/', [GenreController::class, 'index'])->name('genres.index');
-		Route::get('/{id}', [GenreController::class, 'show'])->name('genres.show');
-		Route::post('/', [GenreController::class, 'store'])->name('genres.store');
-		Route::put('/{id}', [GenreController::class, 'update'])->name('genres.update');
-		Route::delete('/{id}', [GenreController::class, 'destroy'])->name('genres.delete');
+		Route::get('/{id}', [GenreController::class, 'show'])->name('genres.show')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::post('/', [GenreController::class, 'store'])->name('genres.store')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::put('/{id}', [GenreController::class, 'update'])->name('genres.update')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::delete('/{id}', [GenreController::class, 'destroy'])->name('genres.delete')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
 	});
 
 	/**
@@ -70,9 +71,10 @@ Route::middleware([JwtMiddleware::class])->group(function () {
 	Route::prefix(NotificationEnum::ROUTE_PREFIX)->group(function () {
 		Route::get('/', [NotificationController::class, 'index'])->name('notifications.index');
 		Route::get('/{id}', [NotificationController::class, 'show'])->name('notifications.show');
-		Route::post('/', [NotificationController::class, 'store'])->name('notifications.store');
-		Route::put('/{id}', [NotificationController::class, 'update'])->name('notifications.update');
-		Route::delete('/{id}', [NotificationController::class, 'destroy'])->name('notifications.delete');
+		Route::get('/{id}/check', [NotificationController::class, 'check'])->name('notifications.check');
+		Route::post('/', [NotificationController::class, 'store'])->name('notifications.store')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::put('/{id}', [NotificationController::class, 'update'])->name('notifications.update')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::delete('/{id}', [NotificationController::class, 'destroy'])->name('notifications.delete')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
 	});
 
 	/**
@@ -81,9 +83,9 @@ Route::middleware([JwtMiddleware::class])->group(function () {
 	Route::prefix(ProductEnum::ROUTE_PREFIX)->group(function () {
 		Route::get('/', [ProductController::class, 'index'])->name('products.index');
 		Route::get('/{id}', [ProductController::class, 'show'])->name('products.show');
-		Route::post('/', [ProductController::class, 'store'])->name('products.store');
-		Route::put('/{id}', [ProductController::class, 'update'])->name('products.update');
-		Route::delete('/{id}', [ProductController::class, 'destroy'])->name('products.delete');
+		Route::post('/', [ProductController::class, 'store'])->name('products.store')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::put('/{id}', [ProductController::class, 'update'])->name('products.update')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::delete('/{id}', [ProductController::class, 'destroy'])->name('products.delete')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
 	});
 
 	/**
@@ -92,7 +94,7 @@ Route::middleware([JwtMiddleware::class])->group(function () {
 	Route::prefix(RatingEnum::ROUTE_PREFIX)->group(function () {
 		Route::get('/', [RatingController::class, 'index'])->name('ratings.index');
 		Route::get('/{id}', [RatingController::class, 'show'])->name('ratings.show');
-		Route::get('/recommend/{id}', [RatingController::class, 'recommend'])->name('ratings.recommend');
+		Route::get('/recommend/{id}', [RatingController::class, 'recommend'])->name('ratings.recommend')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
 		Route::post('/', [RatingController::class, 'store'])->name('ratings.store');
 		Route::put('/{id}', [RatingController::class, 'update'])->name('ratings.update');
 		Route::delete('/{id}', [RatingController::class, 'destroy'])->name('ratings.delete');
@@ -102,12 +104,12 @@ Route::middleware([JwtMiddleware::class])->group(function () {
 	 * SALES
 	 */
 	Route::prefix(SaleEnum::ROUTE_PREFIX)->group(function () {
-		Route::get('/', [SaleController::class, 'index'])->name('sales.index');
-		Route::get('/{id}', [SaleController::class, 'show'])->name('sales.show');
-		Route::get('/report/generate/{period}', [SaleController::class, 'reportGenerate'])->name('sales.reportGenerate');
-		Route::post('/', [SaleController::class, 'store'])->name('sales.store');
-		Route::put('/{id}', [SaleController::class, 'update'])->name('sales.update');
-		Route::delete('/{id}', [SaleController::class, 'destroy'])->name('sales.delete');
+		Route::get('/', [SaleController::class, 'index'])->name('sales.index')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::get('/{id}', [SaleController::class, 'show'])->name('sales.show')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::get('/report/generate/{period}', [SaleController::class, 'reportGenerate'])->name('sales.reportGenerate')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN);
+		Route::post('/', [SaleController::class, 'store'])->name('sales.store')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN . '-' . UserEnum::EMPLOYEE . '');
+		Route::put('/{id}', [SaleController::class, 'update'])->name('sales.update')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN);
+		Route::delete('/{id}', [SaleController::class, 'destroy'])->name('sales.delete')->middleware(RoleMiddleware::class.':' . UserEnum::ADMIN);
 	});
 });
